Reduce the number of messages BCAR sends
After chatting on Discord, we noticed that BCAR is sending an AccountUpdate
for every ear/tail/wing animation which isn't really necessary.
This change adds refresh=false to the InventoryWear call to avoid the
AccountUpdate being sent. This means we do need to call
CharacterRefresh manually so the update shows up locally. We just
send the ChatRoomCharacterItemUpdate instead. This doesn't cause
the server to save the update which make it quite a lot lighter,
and we don't need the updates saved anyway since they're just for
animations.