Fix loading on new accounts
It crashed during loading on new accounts because the settings were
blank and so had no genderDefault so the mod didn't load for people
who'd never used it before.

This just checks that genderDefault is defined. Another fix would be
to call migrate_gender() after the defult settings had been saved,
if you'd prefer that.